feat: operator bracketing rules
This commit introduces operator categories relating to whether
explicit brackets should be added for clarification.

Regarding the left operand, it should be added for non-commutative
operators (sub, div), but not for commutative ones (add, mul, and, or),
as it just increases noise.

On the right side, the associative concat does not get parens, but pow
should, as it might be confusing that it binds right.

Unary minus will not get extra brackets, but # (length) will, if it's
used on a compound identifier.